net/ice/base: improve read retry value calculation
Previous implementation of PHY timestamp retry value was set to a static
value that had no meaning. Change it to calculate it based on a set of
meaningfully named macros, as well as adjust data type to avoid
overflows.

Acked-by: Bruce Richardson <bruce.richardson@intel.com>

+++ drivers/net/ice/base/ice_ptp_hw.c
@@ -4492,7 +4492,7 @@ static int ice_write_phy_reg_e810(struct ice_hw *hw, u32 addr, u32 val)
 static int
 ice_read_phy_tstamp_ll_e810(struct ice_hw *hw, u8 idx, u8 *hi, u32 *lo)
 {
-	u8 i;
+	unsigned int i;
 
 	/* Write TS index to read to the PF register so the FW can read it */
 	wr32(hw, PF_SB_ATQBAL, TS_LL_READ_TS_IDX(idx));
